Basic Fishing Prototype
Here's the basis for the fishing. It's really just the groundwork and is kind of rough and fairly basic, but with some polishing it'll wokr pretty well. Before I can continue on it, I need everyone's opinions on what they'd want out of it and how they'd expect it to work.